fix review10: move tests
